Add testcases exercising general CRL structure
